requirements.txt
appdirs==1.4.3
Beaker==1.11.0
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==0.6.2
oic==1.2.1
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
pycryptodomex==3.9.9
pyjwkest==1.4.2
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
requests==2.25.1
retrying==1.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
